moving all js code into thee js folder
