+++ MANIFEST.in
@@ -2,9 +2,9 @@ include README.md
 
 include setupbase.py
 
-include tsconfig.json
-include package.json
-include webpack.config.js
+include js/tsconfig.json
+include js/package.json
+include js/webpack.config.js
 include jupyter_jsmol/labextension/*.tgz
 
 # Documentation
@@ -16,16 +16,15 @@ prune site
 graft examples
 
 # Tests
-graft tests
-prune tests/build
+graft js/tests
+prune js/tests/build
 
 # Javascript files
 graft jupyter_jsmol/nbextension
-graft src
-graft css
+graft js/src
+graft js/css
 prune **/node_modules
-prune coverage
-prune lib
+prune js/lib
 
 # Patterns to exclude from any directory
 global-exclude *~
