+++ js/.npmignore
