+++ js/package.json
@@ -30,14 +30,14 @@
   },
   "scripts": {
     "build": "npm run build:lib && npm run build:nbextension",
-    "build:labextension": "npm run clean:labextension && mkdirp jupyter_jsmol/labextension && cd jupyter_jsmol/labextension && npm pack ../..",
+    "build:labextension": "npm run clean:labextension && mkdirp ../jupyter_jsmol/labextension && cd ../jupyter_jsmol/labextension && npm pack ../../js",
     "build:lib": "tsc",
     "build:nbextension": "webpack -p",
     "build:all": "npm run build:labextension && npm run build:nbextension",
     "clean": "npm run clean:lib && npm run clean:nbextension",
     "clean:lib": "rimraf lib",
-    "clean:labextension": "rimraf jupyter_jsmol/labextension",
-    "clean:nbextension": "rimraf jupyter_jsmol/nbextension/static/index.js",
+    "clean:labextension": "rimraf ../jupyter_jsmol/labextension",
+    "clean:nbextension": "rimraf ../jupyter_jsmol/nbextension/static/index.js",
     "prepack": "npm run build:lib",
     "test": "npm run test:firefox",
     "test:chrome": "karma start --browsers=Chrome tests/karma.conf.js",
