+++ js/tests/karma.conf.js
