+++ js/tsconfig.json
@@ -17,6 +17,6 @@
   },
   "include": [
     "src/**/*.ts",
-    "src/**/*.tsx",
+    "src/**/*.tsx"
   ]
 }
